requirements.txt
Appium-Python-Client>=2.11.0
selenium>=4.15.0
pytest>=7.4.0
